+++ src/lib/connect-four/board.js
@@ -1,3 +1,5 @@
+const { createCanvas } = require("canvas");
+
 const DIRECTION = {
 	UP: { x: 0, y: -1 },
 	UP_RIGHT: { x: 1, y: -1 },
@@ -7,6 +9,25 @@ const DIRECTION = {
 	DOWN_LEFT: { x: -1, y: 1 },
 	LEFT: { x: -1, y: 0 },
 	UP_LEFT: { x: -1, y: -1 },
+};
+
+const DISCORD_GRAY = "#37393F";
+const PALETTES = {
+	ENOS16: {
+		board: "#58AEEE",
+		one: "#F9D381",
+		two: "#E75952",
+	},
+	PICO8: {
+		board: "#29adff",
+		one: "#ffec27",
+		two: "#ff004d",
+	},
+	WEB: {
+		board: "#3399ff",
+		one: "#ffcc33",
+		two: "#ff3333",
+	},
 };
 
 class Board {
@@ -143,6 +164,108 @@ class Board {
 		}
 
 		return false;
+	}
+
+	/**
+	 * Creates and returns a canvas that represents the current board's state.
+	 * @param {Object} options Optional options to apply to the canvas.
+	 */
+	createCanvas(options) {
+		let padding = 4;
+		let tileSize = 20;
+		let fontSize = Math.floor(tileSize * 0.8);
+		let boardColor = PALETTES.WEB.board;
+		let oneColor = PALETTES.WEB.one;
+		let twoColor = PALETTES.WEB.two;
+
+		if (options !== undefined) {
+			if (options.padding !== undefined) {
+				padding = parseInt(options.padding);
+			}
+
+			if (options.tileSize !== undefined) {
+				tileSize = parseInt(options.tileSize);
+				fontSize = Math.floor(tileSize * 0.8);
+			}
+
+			if (options.palette) {
+				if (options.palette.board !== undefined) {
+					boardColor = options.palette.board;
+				}
+				if (options.palette.one !== undefined) {
+					oneColor = options.palette.one;
+				}
+				if (options.palette.two !== undefined) {
+					twoColor = options.palette.two;
+				}
+			}
+		}
+
+		const canvas = createCanvas(
+			this.columns * tileSize + padding * 2,
+			this.rows * tileSize + padding * 3 + fontSize
+		);
+		const ctx = canvas.getContext("2d");
+
+		// Clear Canvas.
+		ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+		// Draw board.
+		ctx.fillStyle = boardColor;
+		ctx.fillRect(
+			0,
+			padding + fontSize,
+			canvas.width,
+			canvas.height - (padding + fontSize)
+		);
+
+		// Draw text.
+		ctx.font = `bold ${fontSize}px monospace`;
+		ctx.textBaseline = "top";
+		ctx.textAlign = "left";
+
+		for (let i = 0; i < this.columns; i++) {
+			ctx.fillStyle = "#FFFFFF";
+			ctx.fillText(
+				(i + 1).toString(),
+				padding + i * tileSize + tileSize * 0.32,
+				padding - tileSize * 0.1
+			);
+		}
+
+		// Draw pieces.
+		let y = 0;
+		for (let i = 0; i < this.state.length; i++) {
+			const x = i % this.columns;
+
+			switch (this.state[i]) {
+				case 0:
+					ctx.fillStyle = DISCORD_GRAY;
+					break;
+				case 1:
+					ctx.fillStyle = oneColor;
+					break;
+				case 2:
+					ctx.fillStyle = twoColor;
+					break;
+			}
+
+			ctx.beginPath();
+			ctx.arc(
+				padding + x * tileSize + tileSize * 0.5,
+				padding + fontSize + padding + y * tileSize + tileSize * 0.5,
+				(tileSize - padding) * 0.5,
+				0,
+				2 * Math.PI
+			);
+			ctx.fill();
+
+			if (x === this.columns - 1) {
+				y++;
+			}
+		}
+
+		return canvas;
 	}
 }
