+++ src/lib/connect-four/index.js
@@ -1,0 +1,7 @@
+const Board = require("./board.js");
+const CanvasHelper = require("./canvasHelper.js");
+
+module.exports = {
+	Board,
+	CanvasHelper,
+};
