+++ src/lib/obelus/command.js
