+++ src/lib/obelus/pseudoRegex.js
